+++ junit5/src/main/kotlin/fr/xgouchet/elmyr/junit5/params/StringForgeryParamResolver.kt
@@ -17,13 +17,13 @@ internal class StringForgeryParamResolver :
         return if (annotation.regex.isNotEmpty()) {
             forge.aStringMatching(annotation.regex)
         } else when (annotation.type) {
-            StringForgeryType.ALPHABETICAL -> forge.anAlphabeticalString(annotation.case)
-            StringForgeryType.ALPHA_NUMERICAL -> forge.anAlphaNumericalString(annotation.case)
-            StringForgeryType.NUMERICAL -> forge.aNumericalString()
-            StringForgeryType.HEXADECIMAL -> forge.anHexadecimalString(annotation.case)
-            StringForgeryType.WHITESPACE -> forge.aWhitespaceString()
-            StringForgeryType.ASCII -> forge.anAsciiString()
-            StringForgeryType.ASCII_EXTENDED -> forge.anExtendedAsciiString()
+            StringForgeryType.ALPHABETICAL -> forge.anAlphabeticalString(annotation.case, annotation.size)
+            StringForgeryType.ALPHA_NUMERICAL -> forge.anAlphaNumericalString(annotation.case, annotation.size)
+            StringForgeryType.NUMERICAL -> forge.aNumericalString(annotation.size)
+            StringForgeryType.HEXADECIMAL -> forge.anHexadecimalString(annotation.case, annotation.size)
+            StringForgeryType.WHITESPACE -> forge.aWhitespaceString(annotation.size)
+            StringForgeryType.ASCII -> forge.anAsciiString(annotation.size)
+            StringForgeryType.ASCII_EXTENDED -> forge.anExtendedAsciiString(annotation.size)
         }
     }